Refactor the Import- and Export-Runner into separate class to make it easier to test
